pyproject.toml
[project]
name = "cataforge"
version = "0.1.0"
description = "AI Programming: Agent + Skill Workflow Framework"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "pydantic>=2.0",
    "markdown-it-py>=3.0",
    "docker>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
mcp = ["mcp>=1.0"]

[project.scripts]
cataforge = "cataforge.cli.main:cli"

[project.entry-points."cataforge.platforms"]
claude-code = "cataforge.platform.claude_code:ClaudeCodeAdapter"
cursor = "cataforge.platform.cursor:CursorAdapter"
codex = "cataforge.platform.codex:CodexAdapter"
opencode = "cataforge.platform.opencode:OpenCodeAdapter"

[project.entry-points."cataforge.plugins"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cataforge"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
